# Implement `cbrt(A::AbstractMatrix{<:Real})`