--- a/stdlib/LinearAlgebra/src/dense.jl
+++ b/stdlib/LinearAlgebra/src/dense.jl
@@ -907,7 +907,39 @@ end
 sqrt(A::AdjointAbsMat) = adjoint(sqrt(parent(A)))
 sqrt(A::TransposeAbsMat) = transpose(sqrt(parent(A)))
 
-# Cube root of a real-valued matrix
+"""
+    cbrt(A::AbstractMatrix{<:Real})
+
+Computes the real-valued cube root of a real-valued matrix `A`. If `T = cbrt(A)`, then
+we have `T*T*T ≈ A`, see example given below.
+
+If `A` is symmetric, i.e., of type `HermOrSym{<:Real}`, then ([`eigen`](@ref)) is used to
+find the cube root. Otherwise, a specialized version of the p-th root algorithm, given in
+[^S03], is utilized, which exploits the real-valued Schur decomposition ([`schur`](@ref))
+to compute the cube root.
+
+[^S03]
+
+    Matthew I. Smith, "A Schur Algorithm for Computing Matrix pth Roots",
+    SIAM Journal on Matrix Analysis and Applications, vol. 24, 2003, pp. 971–989.
+    [doi:10.1137/S0895479801392697](https://doi.org/10.1137/s0895479801392697)
+
+# Examples
+```jldoctest
+julia> A = randn(2,2)
+2×2 Matrix{Float64}:
+ -1.08459   -0.936176
+ -0.734221  -0.566529
+
+julia> T = cbrt(A)
+2×2 Matrix{Float64}:
+ -0.650849  -0.831283
+ -0.651956  -0.190831
+
+julia> T*T*T ≈ A
+true
+```
+"""
 function cbrt(A::AbstractMatrix{<:Real})
     if checksquare(A) == 0
         return copy(A)