--- a/stdlib/LinearAlgebra/src/symmetric.jl
+++ b/stdlib/LinearAlgebra/src/symmetric.jl
@@ -811,7 +811,7 @@ for func in (:log, :sqrt)
     end
 end
 
-# Cube root for real-valued matrices
+# Cube root of a real-valued symmetric matrix
 function cbrt(A::Symmetric{T}) where {T<:Real}
     F = eigen(A)
     A = F.vectors * Diagonal(cbrt.(F.values)) * F.vectors'