--- a/stdlib/LinearAlgebra/src/triangular.jl
+++ b/stdlib/LinearAlgebra/src/triangular.jl
@@ -2520,11 +2520,41 @@ function _cbrt_blkdiag_1x1_2x2!(A::AbstractMatrix{T}) where {T<:Real}
     # 2x2 and 1x1 blocks
     I2 = findall(x -> !iszero(x), diag(A,-1))
     I1 = setdiff(1:n, vcat(I2, I2.+1))
-    for idx in I2
-        @views _cbrt_2x2!(A[idx:idx+1,idx:idx+1])
-    end
-    for idx in I1
-        A[idx,idx] = cbrt(A[idx,idx])
-    end
+    for idx in I2 @views _cbrt_2x2!(A[idx:idx+1,idx:idx+1]) end
+    for idx in I1 @views A[idx,idx] = cbrt(A[idx,idx]) end
     return A, I2, I1
+end
+
+# Cube root of a quasi triangular matrix (output of Schur decomposition)
+# Reference: Smith, M. I. (2003). A Schur Algorithm for Computing Matrix pth Roots.
+#   SIAM Journal on Matrix Analysis and Applications (Vol. 24, Issue 4, pp. 971–989).
+#   https://doi.org/10.1137/s0895479801392697
+function _cbrt_quasi_triangular!(A::AbstractMatrix{T}) where {T<:BlasReal}
+    m, n = size(A)
+    (m == n) || throw(ArgumentError("_cbrt_quasi_triangular!: Matrix A must be square."))
+    A, I2, I1 = _cbrt_blkdiag_1x1_2x2!(A)
+    sizes = [if i ∈ I1 1 elseif i ∈ I2 2 else 0 end for i=1:n]
+    for k = 1:n-1
+        for i = filter(i -> sizes[i] > 0 && sizes[i+k] > 0, 1:n-k)
+            (i₁, i₂, j₁, j₂, s₁, s₂) = (i, i+sizes[i]-1, i+k, i+k+sizes[i+k]-1, sizes[i], sizes[i+k])
+            Bᵢⱼ⁽⁰⁾ = zeros(T,s₁,s₂)
+            Bᵢⱼ⁽¹⁾ = zeros(T,s₁,s₂)
+            for k₁ = filter(k -> sizes[k] > 0, i+1:i+k-1)
+                k₂ = k₁+sizes[k₁]-1
+                # Bᵢⱼ⁽⁰⁾ = Bᵢⱼ⁽⁰⁾ + A[i₁:i₂,k₁:k₂]*A[k₁:k₂,j₁:j₂]
+                @views BLAS.gemm!('N', 'N', 1.0, A[i₁:i₂,k₁:k₂], A[k₁:k₂,j₁:j₂], 1.0, Bᵢⱼ⁽⁰⁾)
+                # Retreive Rᵢⱼ[i,i+k] as A[i+k,i]'
+                # Bᵢⱼ⁽¹⁾ = Bᵢⱼ⁽¹⁾ + A[i₁:i₂,k₁:k₂]*A[j₁:j₂,k₁:k₂]'
+                @views BLAS.gemm!('N', 'T', 1.0, A[i₁:i₂,k₁:k₂], A[j₁:j₂,k₁:k₂], 1.0, Bᵢⱼ⁽¹⁾)
+            end
+            @views L = kron(I(s₂), A[i₁:i₂,i₁:i₂]^2) + kron(A[j₁:j₂,j₁:j₂]', A[i₁:i₂,i₁:i₂]) + kron((A[j₁:j₂,j₁:j₂]^2)', I(s₁))
+            @views R = A[i₁:i₂,j₁:j₂] - A[i₁:i₂,i₁:i₂]*Bᵢⱼ⁽⁰⁾ - Bᵢⱼ⁽¹⁾
+            @views A[i₁:i₂,j₁:j₂] = reshape(L \ R[:], s₁, s₂)
+            # Store Rᵢⱼ[i,i+k]' in A[i+k,i]
+            @views A[j₁:j₂,i₁:i₂] = A[i₁:i₂,j₁:j₂]'*A[i₁:i₂,i₁:i₂]' + A[j₁:j₂,j₁:j₂]'*A[i₁:i₂,j₁:j₂]' + Bᵢⱼ⁽⁰⁾'
+        end
+    end
+    # Make quasi triangular
+    for j=1:n for i=j+1+(sizes[j]==2):n A[i,j] = 0 end end
+    return A
 end