--- a/stdlib/LinearAlgebra/src/triangular.jl
+++ b/stdlib/LinearAlgebra/src/triangular.jl
@@ -2518,11 +2518,11 @@ function _cbrt_blkdiag_1x1_2x2!(A::AbstractMatrix{T}) where {T<:Real}
     m, n = size(A)
     (m == n) || throw(ArgumentError("_cbrt_blkdiag_1x1_2x2!: Matrix A must be square."))
     # 2x2 and 1x1 blocks
-    I2 = findall(x -> !iszero(x), diag(A,-1))
-    I1 = setdiff(1:n, vcat(I2, I2.+1))
-    for i in I2 @views _cbrt_2x2!(A[i:i+1,i:i+1]) end
-    for i in I1 @views A[i,i] = cbrt(A[i,i]) end
-    return A, I2, I1
+    I₂ = findall(x -> !iszero(x), diag(A,-1))
+    I₁ = setdiff(1:n, vcat(I₂, I₂.+1))
+    for i in I₂ @views _cbrt_2x2!(A[i:i+1,i:i+1]) end
+    for i in I₁ @views A[i,i] = cbrt(A[i,i]) end
+    return A, I₂, I₁
 end
 
 # Cube root of a quasi upper triangular matrix (output of Schur decomposition)
@@ -2532,8 +2532,9 @@ end
 function _cbrt_quasi_triu!(A::AbstractMatrix{T}) where {T<:Real}
     m, n = size(A)
     (m == n) || throw(ArgumentError("_cbrt_quasi_triu!: Matrix A must be square."))
-    A, I2, I1 = _cbrt_blkdiag_1x1_2x2!(A)
-    sizes = [if i ∈ I1 1 elseif i ∈ I2 2 else 0 end for i=1:n]
+    A, I₂, I₁ = _cbrt_blkdiag_1x1_2x2!(A)
+    sizes = [if i ∈ I₁ 1 elseif i ∈ I₂ 2 else 0 end for i=1:n]
+    ID = I(4)
     for k = 1:n-1
         for i = 1:n-k
             if sizes[i] == 0 || sizes[i+k] == 0 continue end
@@ -2542,6 +2543,8 @@ function _cbrt_quasi_triu!(A::AbstractMatrix{T}) where {T<:Real}
             Bᵢⱼ⁽¹⁾ = zeros(T,s₁,s₂)
             L₀ = zeros(T,s₁*s₂,s₁*s₂)
             L₁ = zeros(T,s₁*s₂,s₁*s₂)
+            S₁ = zeros(T,s₁,s₁)
+            S₂ = zeros(T,s₂,s₂)
             for k₁ = i+1:i+k-1
                 if sizes[k₁] == 0 continue end
                 k₂ = k₁+sizes[k₁]-1
@@ -2549,9 +2552,11 @@ function _cbrt_quasi_triu!(A::AbstractMatrix{T}) where {T<:Real}
                 # Retreive Rᵢⱼ[i,i+k] as A[i+k,i]'
                 @views mul!(Bᵢⱼ⁽¹⁾, A[i₁:i₂,k₁:k₂], A[j₁:j₂,k₁:k₂]', 1.0, 1.0)
             end
-            @views kron!(L₀, I(s₂), A[i₁:i₂,i₁:i₂]^2)
+            @views mul!(S₁, A[i₁:i₂,i₁:i₂], A[i₁:i₂,i₁:i₂])
+            @views mul!(S₂, A[j₁:j₂,j₁:j₂], A[j₁:j₂,j₁:j₂])
+            @views kron!(L₀, ID[1:s₂,1:s₂], S₁)
             @views L₀ .+= kron!(L₁, A[j₁:j₂,j₁:j₂]', A[i₁:i₂,i₁:i₂])
-            @views L₀ .+= kron!(L₁, (A[j₁:j₂,j₁:j₂]^2)', I(s₁))
+            @views L₀ .+= kron!(L₁, S₂', ID[1:s₁,1:s₁])
             @views mul!(A[i₁:i₂,j₁:j₂], A[i₁:i₂,i₁:i₂], Bᵢⱼ⁽⁰⁾, -1.0, 1.0)
             @views A[i₁:i₂,j₁:j₂] .-= Bᵢⱼ⁽¹⁾
             @views ldiv!(lu!(L₀), A[i₁:i₂,j₁:j₂][:])