--- a/stdlib/LinearAlgebra/src/triangular.jl
+++ b/stdlib/LinearAlgebra/src/triangular.jl
@@ -2534,17 +2534,19 @@ function _cbrt_quasi_triu!(A::AbstractMatrix{T}) where {T<:Real}
     (m == n) || throw(ArgumentError("_cbrt_quasi_triu!: Matrix A must be square."))
     A, I₂, I₁ = _cbrt_blkdiag_1x1_2x2!(A)
     sizes = [if i ∈ I₁ 1 elseif i ∈ I₂ 2 else 0 end for i=1:n]
-    ID = I(4)
+    Δ = I(4)
+    L₀ᴹ = zeros(T,4,4)
+    L₁ᴹ = zeros(T,4,4)
     for k = 1:n-1
         for i = 1:n-k
             if sizes[i] == 0 || sizes[i+k] == 0 continue end
             i₁, i₂, j₁, j₂, s₁, s₂ = i, i+sizes[i]-1, i+k, i+k+sizes[i+k]-1, sizes[i], sizes[i+k]
-            Bᵢⱼ⁽⁰⁾ = zeros(T,s₁,s₂)
-            Bᵢⱼ⁽¹⁾ = zeros(T,s₁,s₂)
-            L₀ = zeros(T,s₁*s₂,s₁*s₂)
-            L₁ = zeros(T,s₁*s₂,s₁*s₂)
             S₁ = zeros(T,s₁,s₁)
             S₂ = zeros(T,s₂,s₂)
+            Bᵢⱼ⁽⁰⁾ = zeros(T,s₁,s₂)
+            Bᵢⱼ⁽¹⁾ = zeros(T,s₁,s₂)
+            @views L₀ = L₀ᴹ[1:s₁*s₂,1:s₁*s₂]
+            @views L₁ = L₁ᴹ[1:s₁*s₂,1:s₁*s₂]
             for k₁ = i+1:i+k-1
                 if sizes[k₁] == 0 continue end
                 k₂ = k₁+sizes[k₁]-1
@@ -2554,16 +2556,16 @@ function _cbrt_quasi_triu!(A::AbstractMatrix{T}) where {T<:Real}
             end
             @views mul!(S₁, A[i₁:i₂,i₁:i₂], A[i₁:i₂,i₁:i₂])
             @views mul!(S₂, A[j₁:j₂,j₁:j₂], A[j₁:j₂,j₁:j₂])
-            @views kron!(L₀, ID[1:s₂,1:s₂], S₁)
+            @views kron!(L₀, Δ[1:s₂,1:s₂], S₁)
             @views L₀ .+= kron!(L₁, A[j₁:j₂,j₁:j₂]', A[i₁:i₂,i₁:i₂])
-            @views L₀ .+= kron!(L₁, S₂', ID[1:s₁,1:s₁])
+            @views L₀ .+= kron!(L₁, S₂', Δ[1:s₁,1:s₁])
             @views mul!(A[i₁:i₂,j₁:j₂], A[i₁:i₂,i₁:i₂], Bᵢⱼ⁽⁰⁾, -1.0, 1.0)
             @views A[i₁:i₂,j₁:j₂] .-= Bᵢⱼ⁽¹⁾
             @views ldiv!(lu!(L₀), A[i₁:i₂,j₁:j₂][:])
             # Store Rᵢⱼ[i,i+k]' in A[i+k,i]
-            @views mul!(A[j₁:j₂,i₁:i₂], A[i₁:i₂,j₁:j₂]', A[i₁:i₂,i₁:i₂]', 1.0, 0)
-            @views mul!(A[j₁:j₂,i₁:i₂], A[j₁:j₂,j₁:j₂]', A[i₁:i₂,j₁:j₂]', 1.0, 1.0)
-            @views A[j₁:j₂,i₁:i₂] .+= Bᵢⱼ⁽⁰⁾'
+            @views mul!(Bᵢⱼ⁽⁰⁾, A[i₁:i₂,i₁:i₂], A[i₁:i₂,j₁:j₂], 1.0, 1.0)
+            @views mul!(Bᵢⱼ⁽⁰⁾, A[i₁:i₂,j₁:j₂], A[j₁:j₂,j₁:j₂], 1.0, 1.0)
+            @views A[j₁:j₂,i₁:i₂] .= Bᵢⱼ⁽⁰⁾'
         end
     end
     # Make quasi triangular