… NEWS.md
… VERSION
… doc/src/base/math.md
… stdlib/LinearAlgebra/docs/src/index.md
… stdlib/LinearAlgebra/test/dense.jl