This PR resolves #47513.
~~Baseline: 1.10.0-alpha1~~ (rebased to `master`)

# Summary of the Implementation
1. ~~`cbrt(A::AbstractMatrix)` is delegated to `A^(1//3)`.~~ (Method removed.)
2. For real-valued `Diagonal` and `Symmetric` matrices, straightforward implementations based on scalar `cbrt` and `eigen` are provided.
3. For real-valued arbitrary matrices, a `schur` decomposition and [1, Algorithm 4.3]-based algorithm (`_cbrt_quasi_triu!`) is provided.
4. For real-valued upper and lower triangular matrices, straightforward solutions based on `_cbrt_quasi_triu!` are provided.

Note: See comment below for details on the algorithm.

## References
|  |   |
| ---- | ---- |
| [1]&nbsp;&nbsp; | Smith, M. I. (2003). A Schur Algorithm for Computing Matrix pth Roots. In SIAM Journal on Matrix Analysis and Applications (Vol. 24, Issue 4, pp. 971–989). https://doi.org/10.1137/s0895479801392697 |

# Pending Items
- [X] ~~Memory and MIPS optimisations (trade-off) for `_cbrt_quasi_triu!`.~~

# Tests
File: `stdlib/LinearAlgebra/test/dense.jl`
Testset: `@testset "cbrt(A::AbstractMatrix{T})"` 

```julia
Test Summary:              | Pass  Total  Time
cbrt(A::AbstractMatrix{T}) |   17     17  4.1s
```

# Memory and MIPS
```julia
julia> A = randn(1000,1000) ;
julia> @btime cbrt(A) ;
  836.035 ms (132377 allocations: 42.71 MiB)
julia> @btime A^(1//3) ;
  6.878 s (180 allocations: 1.07 GiB)
julia> @btime sqrt(A) ;
  888.409 ms (24 allocations: 91.86 MiB)
```

- Target was to get lower than `sqrt` wrt MIPS and memory consumption for large $N,$ which is done now 🎉🎉🎉