@aravindh-krishnamoorthy 2023-07-24
# Algorithm overview: Cube root of a quasi triangular matrix

Cube root of a quasi triangular matrix is the crucial part of this PR. The implementation consists of three functions implemented in `stdlib/LinearAlgebra/src/triangular.jl`. See "Files Changed" tab for details.

- [X] `_cbrt_2x2!`: Cube root of a 2x2 real-valued matrix with complex conjugate eigenvalues and equal diagonal values. [Permalink to method.](https://github.com/aravindh-krishnamoorthy/julia/blob/cbrt/stdlib/LinearAlgebra/src/triangular.jl#L2495)
- [X] `_cbrt_blkdiag_1x1_2x2!`: Cube root of a block diagonal matrix consisting of $1\times1$ and $2\times2$ blocks (output of Schur decomposition) -- first step of Algorithm 4.3 in [2]. [Permalink to method.](https://github.com/aravindh-krishnamoorthy/julia/blob/cbrt/stdlib/LinearAlgebra/src/triangular.jl#L2516)
- [X] `_cbrt_quasi_triu!`: Cube root of a quasi-triangular matrix, i.e., remainder of Algorithm 4.3 in [2]. [Permalink to method.](https://github.com/aravindh-krishnamoorthy/julia/blob/cbrt/stdlib/LinearAlgebra/src/triangular.jl#L2528)

A first version of the `_cbrt_quasi_triu!` function for real-valued quasi-triangular matrices per Algorithm 4.3 in [1] (see PR text), can be accessed via the permalinks above or via the "Files Changed" tab. The algorithm can be optimised further for both execution time and memory usage. Please see my implementation notes below. An example run is as follows:

```julia
julia> A = randn(1000,1000) ;
julia> T = schur(A).T ;
julia> U = LinearAlgebra._cbrt_quasi_triu!(copy(T)) ;
julia> typeof(U)
Matrix{Float64} (alias for Array{Float64, 2})
julia> norm(U*U*U-T)
1.9153193181259216e-11
julia> @btime U = LinearAlgebra._cbrt_quasi_triu!(T₀) setup=(T₀ = copy(T)) evals=1 samples=100 ;
  2.050 s (3650325 allocations: 865.11 MiB)

julia> V = T^(1/3) ;
julia> typeof(V)
Matrix{ComplexF64} (alias for Array{Complex{Float64}, 2})
julia> norm(V*V*V-T)
2.787726919534373e-10
julia> @btime V = T^(1/3) evals=1 samples=100 ;
  4.754 s (155 allocations: 900.61 MiB)
```

### Implementation Notes:
- In principle, the algorithm is **in place** because the intermediate matrices $\boldsymbol{R}_{ij}$ are stored in the lower-triangular parts.
- The present memory overheads are due to allocation of intermediate values. These can likely be optimised away to some extent.
- The current version is for `<:BlasFloat` and uses `gemm` in two places. Perhaps this part can also be rewritten in Julia efficiently. See pending items in the PR text.
- Although [2] reports $p$-th root of a matrix, the present implementation is unfortunately? optimised for cube roots. A general solution will require a lot more memory and operation, but can return real-valued matrices whenever the $p$-th root of the block diagonal matrix (analogous to `_cbrt_blkdiag_1x1_2x2!`) is real valued.