@aravindh-krishnamoorthy 2023-07-26
Hello @dkarrasch, at your convenience, could you please suggest a solution for the following. I have a $4\times 4$ chunk of memory in
```julia
A = zeros(T, 4, 4)
```
What would be a good way to get views into it with sizes $\\{1,2,4\\} \times \\{1,2,4\\},$ i.e., any combinations of two of three numbers (e.g., $1\times 2,$ $4\times 1,$ $2\times 4,$ $4\times 4$), such that:
- The view is zero initialized,
- The view is contiguous (column major) and can be passed to `mul!` (or LAPACK),
- The view is (almost nearly) overhead free,
- Bounds check works throughout.

Such a solution will help me reduce the memory usage of `_cbrt_quasi_triu!` even further. As you see now, the fastest solution seems to be to allocate these small matrices every loop:
```julia
S₁ = zeros(T,s₁,s₁)
S₂ = zeros(T,s₂,s₂)
Bᵢⱼ⁽⁰⁾ = zeros(T,s₁,s₂)
Bᵢⱼ⁽¹⁾ = zeros(T,s₁,s₂)
```
GC doesn't seem to run at the end of every loop so they seem to accumulate. Maximum memory utilization is about $70$ MiB, so I've left it as is for now.