@stevengj 2023-08-01
Looks good, thanks for working on this!  You should add an item to NEWS.md