@barucden 2023-08-02
It might be a good idea to add similar docs as we [have for `sqrt`](https://github.com/JuliaLang/julia/blob/3c64bc3dbeb5161a27f7df33c2e3ee3df1c32e5a/stdlib/LinearAlgebra/src/dense.jl#L829-L875)?