@barucden 2023-08-03
I've been under the impression that a documentation is (or should be) an integral part of any exported functionality. So I'd say it should be part of this PR. BUT I am not a maintainer! I just noticed that the documentation is missing here, so I thought I should point it out. In any case, there is no special team working on the documentation.