@aravindh-krishnamoorthy 2023-08-05
> https://github.com/JuliaLang/julia/blob/2c45e3ba9a2cf801d5b8da433552541c196ab54b/doc/src/base/math.md#L161
> 
> should be changed to `Base.Math.cbrt(::Number)` so it only documents the scalar version, similar to `sqrt` a few lines above.
> And a reference to `cbrt(::AbstractMatrix)` should be added around here so that it shows up in the LinearAlgebra manual:
> 
> https://github.com/JuliaLang/julia/blob/2c45e3ba9a2cf801d5b8da433552541c196ab54b/stdlib/LinearAlgebra/docs/src/index.md?plain=1#L518

Thank you for the tip. This is now done and the docs build well on my PC.